WIP: Add a first rough visualization of per-thread event timelines
We now record all events per thread and then put that data into a
model. A new TimeLineDelegate visualizes the events for every thread
in a simplistic bar chart, i.e. sample cost over time.

This is already quite interesting as it visualizes the contributions
of individual threads to the overall cost. But there are still
some rough edges here:

- no way to actually filter per thread
- no way to actually filter based on time
- time line looks bad

Fixes: https://github.com/KDAB/hotspot/issues/7